refactor: remove unnecessary code
fix: relayer key creation logic
