+++ relayer/query.go
@@ -1,13 +1,17 @@
 package relayer
 
 import (
+	"math/big"
+
 	"github.com/dymensionxyz/roller/cmd/consts"
 	"github.com/dymensionxyz/roller/cmd/utils"
 	"github.com/dymensionxyz/roller/sequencer"
 	"github.com/dymensionxyz/roller/utils/config"
 )
 
-func GetRolRlyAccData(cfg config.RollappConfig) (*utils.AccountData, error) {
+var oneDayRelayPrice = big.NewInt(1)
+
+func getRolRlyAccData(cfg config.RollappConfig) (*utils.AccountData, error) {
 	RollappRlyAddr, err := utils.GetRelayerAddress(cfg.Home, cfg.RollappID)
 	seq := sequencer.GetInstance(cfg)
 	if err != nil {
@@ -31,25 +35,7 @@ func GetRolRlyAccData(cfg config.RollappConfig) (*utils.AccountData, error) {
 	}, nil
 }
 
-func GetRelayerAccountsData(cfg config.RollappConfig) ([]utils.AccountData, error) {
-	data := []utils.AccountData{}
-
-	rollappRlyAcc, err := GetRolRlyAccData(cfg)
-	if err != nil {
-		return nil, err
-	}
-
-	data = append(data, *rollappRlyAcc)
-	hubRlyAcc, err := GetHubRlyAccData(cfg)
-	if err != nil {
-		return nil, err
-	}
-
-	data = append(data, *hubRlyAcc)
-	return data, nil
-}
-
-func GetHubRlyAccData(cfg config.RollappConfig) (*utils.AccountData, error) {
+func getHubRlyAccData(cfg config.RollappConfig) (*utils.AccountData, error) {
 	HubRlyAddr, err := utils.GetRelayerAddress(cfg.Home, cfg.HubData.ID)
 	if err != nil {
 		return nil, err
@@ -70,4 +56,50 @@ func GetHubRlyAccData(cfg config.RollappConfig) (*utils.AccountData, error) {
 		Address: HubRlyAddr,
 		Balance: HubRlyBalance,
 	}, nil
+}
+
+func GetRelayerAccountsData(cfg config.RollappConfig) ([]utils.AccountData, error) {
+	data := []utils.AccountData{}
+
+	rollappRlyAcc, err := getRolRlyAccData(cfg)
+	if err != nil {
+		return nil, err
+	}
+
+	data = append(data, *rollappRlyAcc)
+	hubRlyAcc, err := getHubRlyAccData(cfg)
+	if err != nil {
+		return nil, err
+	}
+
+	data = append(data, *hubRlyAcc)
+	return data, nil
+}
+
+func GetRelayerInsufficientBalances(
+	config config.RollappConfig,
+) ([]utils.NotFundedAddressData, error) {
+	var insufficientBalances []utils.NotFundedAddressData
+
+	accData, err := GetRelayerAccountsData(config)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, acc := range accData {
+		if acc.Balance.Amount.Cmp(oneDayRelayPrice) < 0 {
+			insufficientBalances = append(
+				insufficientBalances, utils.NotFundedAddressData{
+					KeyName:         consts.KeysIds.RollappRelayer,
+					Address:         acc.Address,
+					CurrentBalance:  acc.Balance.Amount,
+					RequiredBalance: oneDayRelayPrice,
+					Denom:           config.Denom,
+					Network:         config.RollappID,
+				},
+			)
+		}
+	}
+
+	return insufficientBalances, nil
 }
